Script changes

+++ src/server/integration/py/teste.py
@@ -115,9 +115,6 @@ def permanentDeleteFile(gDriveService, fileGdriveId):
 	except:
 		pass
 
-
-
-
 def getImageFromLandsat(dtStart, dtEnd, path, row, year):
 	imgResult = 0;
 	#dtStart = str(year)+'-01-01';
@@ -206,60 +203,60 @@ def downloadLandsatFromEE(Id, longitude, latitude, startYear, endYear, startChuv
 
 				landsatBands, dataAcquired, spacecraft, img = getImageFromLandsat(dtStart, dtEnd, path, row, year);
 				
-				try:					
+				if spacecraft != '':
+					try:					
 
-					lon = str("%.4f" % longitude);
-					lat = str("%.4f" % latitude);				
+						lon = str("%.4f" % longitude);
+						lat = str("%.4f" % latitude);				
 
-					filename = Id+"_"+dataAcquired+"_"+spacecraft.lower() + '_' + x['type'];
+						filename = Id+"_"+dataAcquired+"_"+spacecraft.lower() + '_' + x['type'];
 
-					imgResult = img.clip(bufferArea);
-					img = imgResult.visualize(bands=landsatBands)
+						imgResult = img.clip(bufferArea);
+						img = imgResult.visualize(bands=landsatBands)
 
-					coordList = bufferArea.coordinates().getInfo();
-					region = [coordList[0][0], coordList[0][1], coordList[0][2], coordList[0][3]];
-					taskConfig['region'] = [coordList[0][0], coordList[0][1], coordList[0][2], coordList[0][3]]
+						coordList = bufferArea.coordinates().getInfo();
+						region = [coordList[0][0], coordList[0][1], coordList[0][2], coordList[0][3]];
+						taskConfig['region'] = [coordList[0][0], coordList[0][1], coordList[0][2], coordList[0][3]]
 
-					task = ee.batch.Export.image.toDrive(image=img, description=filename, folder=folder, region=region, scale=30)
-					task.start()
-
-					status = task.status()	
-					taskStatus = status['state']
-
-					fileGdriveId=None;
-					geofilename = filename
-					
-					filename = filename + '.tif'
-
-										
-					while fileGdriveId == None or taskStatus not in (ee.batch.Task.State.FAILED, ee.batch.Task.State.COMPLETED, ee.batch.Task.State.CANCELLED):
-						
-						time.sleep(GDRIVE_SLEEP_TIME)
-						fileGdriveId = getFileId(gDriveService, filename);
+						task = ee.batch.Export.image.toDrive(image=img, description=filename, folder=folder, region=region, scale=30)
+						task.start()
 
 						status = task.status()	
 						taskStatus = status['state']
-						taskDescri = task.config['description']
-						
-					if fileGdriveId != None:
-						downloadFile(gDriveService, fileGdriveId, filename);
-						permanentDeleteFile(gDriveService, fileGdriveId);
-						os.system("gdalwarp -t_srs EPSG:4326 "+ filename+" "+geofilename+"-geo.tif"+" "+"&>"+" "+"/dev/null")
 
-						imageFiles.append(str(geofilename+"-geo"));
+						fileGdriveId=None;
+						geofilename = filename
 						
-						try:
-							os.remove(filename);
-						except:
-							traceback.print_exc();
+						filename = filename + '.tif'
+		
+						while fileGdriveId == None and taskStatus not in (ee.batch.Task.State.FAILED, ee.batch.Task.State.COMPLETED, ee.batch.Task.State.CANCELLED):
+							
+							time.sleep(GDRIVE_SLEEP_TIME)
+							fileGdriveId = getFileId(gDriveService, filename);
+
+							status = task.status()	
+							taskStatus = status['state']
+							taskDescri = task.config['description']
+							
+						if fileGdriveId != None:
+							downloadFile(gDriveService, fileGdriveId, filename);
+							permanentDeleteFile(gDriveService, fileGdriveId);
+							os.system("gdalwarp -t_srs EPSG:4326 "+ filename+" "+geofilename+"-geo.tif"+" "+"&>"+" "+"/dev/null")
+
+							imageFiles.append(str(geofilename+"-geo"));
+							
+							try:
+								os.remove(filename);
+							except:
+								traceback.print_exc();
+								pass
+
+						else:
 							pass
-
-					else:
-						pass
-					
-				except:
-					traceback.print_exc();
-					pass;
+						
+					except:
+						traceback.print_exc();
+						pass;
 
 	return imageFiles;
